+++ EDutils/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 ## 1.2dev
+### Tue 18 Jul 14:05:19 BST 2023
+- added path option to import pets
+- added manual integration to dials
+
 ### Fri  9 Jun 17:51:51 BST 2023
 - quick update to the smv reader